src: reorganize source files
